resolve out of sync problem
